% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{amp.acro}
\alias{amp.acro}
\title{Checks the validity of user inputs and creates formula and modifies dataframe}
\usage{
amp.acro(
  time_col,
  n_components = 1,
  group,
  .data,
  .formula,
  period = 12,
  .quietly = TRUE,
  .amp.acro_ind = -1,
  .data_prefix = "main_"
)
}
\arguments{
\item{time_col}{a column of time values in the dataframe}

\item{n_components}{number of components in the model}

\item{group}{a vector of the names of group factors. The levels of each factor should be ordered, with the first level of each factor being the reference level}

\item{.data}{the dataframe from the original cosinor.glmm() function}

\item{.formula}{the formula from the original cosinor.glmm() function}

\item{period}{the period of each component}

\item{.quietly}{controls whether messages from amp.acro are displayed. TRUE by default}
}
\value{
updated dataframe and formula to then be processed by data_processor()
}
\description{
Checks the validity of user inputs and creates formula and modifies dataframe
}
