% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_acro.R
\name{amp_acro}
\alias{amp_acro}
\title{Checks the validity of user inputs and creates formula and modifies dataframe}
\usage{
amp_acro(time_col, n_components = 1, group, period, ...)
}
\arguments{
\item{time_col}{a column of time values in the dataframe}

\item{n_components}{number of components in the model}

\item{group}{a vector of the names of group factors. The levels of each factor should be ordered, with the first level of each factor being the reference level}

\item{period}{the period of each component}

\item{...}{extra arguments controlled by GLMMcosinor.}
}
\value{
updated dataframe and formula to then be processed by data_processor()
}
\description{
Checks the validity of user inputs and creates formula and modifies dataframe
}
