% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.glmm.R
\name{cosinor.glmm}
\alias{cosinor.glmm}
\title{Fit cosinor model with \code{{glmmTMB}}}
\usage{
cosinor.glmm(
  formula,
  data,
  family = stats::gaussian(),
  quietly = TRUE,
  dispformula = ~1,
  ziformula = ~0,
  ...
)
}
\arguments{
\item{formula}{Formula specifying the model. Indicate the time variable with
\code{time()} and covariate effects on the amplitude and acrophase with
\code{amp.acro()}. See details for more information.}

\item{data}{Data frame where variable can be found}

\item{family}{a family function, see \code{?family} and \code{?glmmTMB::nbinom2}}

\item{quietly}{controls whether messages from amp.acro are displayed. TRUE by default}

\item{dispformula}{Formula specifying a dispersion model (optional). Use the same format as the main formula}

\item{ziformula}{Formula specifying a zero-inflation model (optional). Use the same format as the main formula}

\item{...}{optional additional arguments passed to glmmTMB::glmmTMB()}
}
\description{
Given an outcome and time variable, fit the cosinor model with optional
covariate effects.
}
\details{
This defines special functions that are used in the formula to
indicate the time variable and which covariates effect the amplitude. To
indicate the time variable wrap the name of it in the function
\code{time()}. To indicate a variable which affects the
acrophase/amplitude, wrap the name in \code{amp.acro()}. This will then do
all the transformations for you. See examples for usage.
}
\examples{

cosinor.glmm(Y ~ X + amp.acro(time,
                             n_components = 3,
                             group = "X",
                             period = c(12, 8, 9)), data = vitamind)

}
\references{
Tong, YL. Parameter Estimation in Studying Circadian Rhythms, Biometrics (1976). 32(1):85--94.
}
