% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processor.R
\name{data_processor}
\alias{data_processor}
\title{Title}
\usage{
data_processor(
  newdata,
  newformula,
  vec_sss,
  vec_rrr,
  n_components,
  group_stats,
  group,
  group_check,
  period,
  family,
  Terms,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
