% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{data_processor_plot}
\alias{data_processor_plot}
\title{Process data for autoplot}
\usage{
data_processor_plot(x, newdata, x_str, ranef_plot = NULL)
}
\arguments{
\item{x}{cglmm object.}

\item{newdata}{new dataset (if used).}

\item{x_str}{A \code{character} vector naming variable(s) to be plotted.
Default has no value and plots all groups.}

\item{ranef_plot}{Specify the random effects variables that you wish to plot.
If not specified, only the fixed effects will be visualized.}
}
\value{
a \code{data.frame}.
}
\description{
Process data for autoplot
}
\examples{
object <- cglmm(
  vit_d ~ amp_acro(time_col = time, group = "X", period = 12),
  data = vitamind
)
data_processor_plot(
  object,
  vitamind[ "time"],
  x_str = "X"
)
}
