% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_acro.R
\name{.amp_acro}
\alias{.amp_acro}
\title{Checks the validity of user inputs and creates formula and modifies dataframe}
\usage{
.amp_acro(
  time_col,
  n_components = 1,
  group,
  period,
  .data,
  .formula,
  .quietly = TRUE,
  .amp_acro_ind = -1,
  .data_prefix = "main_",
  .env
)
}
\arguments{
\item{time_col}{a column of time values in the dataframe. Can be a string or an object}

\item{n_components}{number of components in the model}

\item{group}{a vector of the names of group factors. The levels of each factor should be ordered, with the first level of each factor being the reference level.}

\item{period}{the period of each component}

\item{.data}{the dataframe from the original cosinor.glmm() function}

\item{.formula}{the formula from the original cosinor.glmm() function}

\item{.quietly}{controls whether messages from amp_acro are displayed. TRUE by default}

\item{.amp_acro_ind}{the index of the portion of the formula containing
amp_acro. -1 for main formula (default), 0 for zi or disp formulae.}

\item{.data_prefix}{prefix for columns to be added in the new dataframe.
Defaults to "main_".}

\item{.env}{The environment in which to evaluate column names on the data being passed.}
}
\value{
updated dataframe and formula to then be processed by data_processor()
}
\description{
Checks the validity of user inputs and creates formula and modifies dataframe
}
