% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cosinor.glmm.R
\name{ggplot.cosinor.glmm.polar}
\alias{ggplot.cosinor.glmm.polar}
\title{Generates a polar plot with elliptical confidence intervals}
\usage{
\method{ggplot}{cosinor.glmm.polar}(
  object,
  x_str,
  contour_interval,
  grid_angle_segments = 8,
  quietly = TRUE,
  radial_units = "radians",
  clockwise = FALSE,
  text_size = 3,
  text_opacity = 0.5,
  component_index = 1,
  make_cowplot = TRUE,
  show_polar_grid = TRUE,
  circle_linetype = "dotted",
  ellipse_opacity = 0.3
)
}
\description{
Generates a polar plot with elliptical confidence intervals
}
