% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMMcosinor-package.R
\docType{package}
\name{GLMMcosinor-package}
\alias{GLMMcosinor}
\alias{GLMMcosinor-package}
\title{GLMMcosinor: Fit a cosinor model using a generalised mixed modelling framework}
\description{
To learn more about how to use GLMMcosinor, see the vignettes at the
\href{https://rwparsons.github.io/GLMMcosinor/}{website} or using
the following code:

\code{browseVignettes(package = "GLMMcosinor")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RWParsons/GLMMcosinor}
  \item \url{https://rwparsons.github.io/GLMMcosinor/}
  \item Report bugs at \url{https://github.com/RWParsons/GLMMcosinor/issues}
}

}
\author{
\strong{Maintainer}: Rex Parsons \email{rex.parsons94@gmail.com} (\href{https://orcid.org/0000-0002-6053-8174}{ORCID})

Authors:
\itemize{
  \item Oliver Jayasinghe
}

}
\keyword{internal}
