% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMMcosinor-package.R
\name{GLMMCosinor}
\alias{GLMMCosinor}
\title{GLMMcosinor: A package for fitting generalized linear mixed-effects models with cosinor terms}
\description{
This package provides functions for fitting generalized linear mixed-effects models with cosinor terms. Cosinor terms can be used to model periodic variations in time series data, and are particularly useful in chronobiology and sleep research. The package uses the flexible GLMMtmb framework that allows users to specify a wide range of data families, and provides tools for visualization.
}
\keyword{GLMM,}
\keyword{chronobiology,}
\keyword{cosinor,}
\keyword{internal}
\keyword{mixed-effects}
\keyword{models,}
\keyword{research}
\keyword{series,}
\keyword{sleep}
\keyword{time}
