% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cosinor.glmm.R
\name{predict.cosinor.glmm}
\alias{predict.cosinor.glmm}
\title{Predict from a cosinor model}
\usage{
\method{predict}{cosinor.glmm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cosinor.glmm}}

\item{...}{other arguments}

\item{newdata}{Optional new data}
}
\description{
Given a time variable and optional covariates, generate predicted values from
a cosinor fit. Default prediction is the mean value, optionally can predict
at a given month
}
\examples{

fit <- cosinor.glmm(Y ~ time(time) + X + amp.acro(X), data = vitamind)
predict(fit)

}
