% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.cglmmSummary}
\alias{print.cglmmSummary}
\title{Print the summary of a cosinor model}
\usage{
\method{print}{cglmmSummary}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{An object of class \code{cglmmSummary}}

\item{digits}{Controls the number of digits displayed in the summary output}

\item{...}{Currently unused}
}
\value{
\code{print} returns \code{x} invisibly.
}
\description{
Print the summary of a cosinor model
}
\examples{

fit <- cglmm(Y ~ X + amp_acro(time,
  group = "X",
  n_components = 1,
  period = 12
), data = vitamind)
summary(fit)

}
