% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cosinor.R
\name{print.sub_test_cosinor}
\alias{print.sub_test_cosinor}
\title{Print test of model}
\usage{
\method{print}{sub_test_cosinor}(x, ...)
}
\arguments{
\item{x}{A \code{sub_test_cosinor} object.}

\item{...}{Additional, ignored arguments.}
}
\value{
\code{print(x)} returns \code{x} invisibly.
}
\description{
Print test of model
}
