% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cosinor.glmm.R
\name{print.summary.cosinor.glmm}
\alias{print.summary.cosinor.glmm}
\title{Print the summary of a cosinor model}
\usage{
\method{print}{summary.cosinor.glmm}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{An object of class \code{summary.cosinor.glmm}}

\item{digits}{Controls the number of digits displayed in the summary output}

\item{...}{Currently unused}
}
\value{
\code{print} returns \code{x} invisibly.
}
\description{
Print the summary of a cosinor model
}
\examples{

fit <- cosinor.glmm(Y ~ X + amp.acro(time,
  group = "X",
  n_components = 1,
  period = 12
), data = vitamind)
summary(fit)

}
