% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cosinor.R
\name{print.test_cosinor}
\alias{print.test_cosinor}
\title{Print results of test of cosinor model}
\usage{
\method{print}{test_cosinor}(x, ...)
}
\arguments{
\item{x}{A \code{test_cosinor} object.}

\item{...}{Arguments passed to \code{print}}
}
\value{
\code{print(x)} returns \code{x} invisibly.
}
\description{
Print results of test of cosinor model
}
