% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMMcosinor-package.R, R/autoplot.R, R/sigma.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{bell}
\alias{autoplot}
\alias{sigma}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{glmmTMB}{\code{\link[glmmTMB:nbinom2]{bell}}}

  \item{stats}{\code{\link[stats]{sigma}}}
}}

