% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cosinor.glmm.R
\name{simulate_cosinor.glmm}
\alias{simulate_cosinor.glmm}
\title{Simulate data from a cosinor model}
\usage{
simulate_cosinor.glmm(
  n,
  amp = 5,
  acro = 10,
  mesor = 0,
  beta.amp = 2,
  beta.acr = 0,
  beta.mesor = 0,
  period = 12,
  dist = c("poisson", "binomial", "gamma", "gaussian")
)
}
\arguments{
\item{n}{Sample size}

\item{beta.amp}{Effect on the amplitude}

\item{beta.mean}{Effect on the mean (intercept)}

\item{beta.acro}{Effect on the acrophase}
}
\description{
This function simulates data from a cosinor model with a single covariate,
where the time scale is month, and optionally
allows for single covariate effects on the mean,
amplitude, and acrophase.
}
