% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cosinor.R
\name{simulate_cosinor}
\alias{simulate_cosinor}
\title{Simulate data from a cosinor model}
\usage{
simulate_cosinor(
  n,
  mesor = 5,
  amp = 10,
  acro = 0,
  beta.mesor = 2,
  beta.amp = 0,
  beta.acro = 0,
  period = 12,
  dist = c("gaussian", "poisson")
)
}
\arguments{
\item{n}{Sample size}

\item{beta.amp}{Effect on the amplitude}

\item{beta.acro}{Effect on the acrophase}

\item{beta.mean}{Effect on the mean (intercept)}
}
\description{
This function simulates data from a cosinor model with a single covariate,
where the time scale is month, and optionally
allows for single covariate effects on the mean,
amplitude, and acrophase.
}
