% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_plot-sub-polar.R
\name{sub_ggplot.cglmm.polar}
\alias{sub_ggplot.cglmm.polar}
\title{create a (sub) polar plot (iterated over for each component)}
\usage{
sub_ggplot.cglmm.polar(
  comp,
  x,
  sum,
  direction,
  offset,
  radial_units,
  grid_angle_segments,
  n_breaks,
  zoom,
  circle_linetype,
  text_size,
  text_opacity,
  ellipse_opacity,
  overlay_parameter_info,
  fill_colors_check,
  xlims_check,
  xlims,
  ylims_check,
  ylims,
  quietly,
  overlay_start,
  fill_colors,
  zoom_origin,
  ranef_plot = NULL,
  ranef_ci = NULL
)
}
\arguments{
\item{comp}{Component.}

\item{x}{cglmm object.}

\item{sum}{Model summary.}

\item{direction}{\code{ifelse(clockwise, -1, 1)}.}

\item{offset}{Applied to start position of phase angle.}

\item{radial_units}{Units for phase.}

\item{grid_angle_segments}{Number of segments.}

\item{n_breaks}{Number of breaks for amplitude.}

\item{zoom}{Whether or not it is zoomed in.}

\item{circle_linetype}{linetype for concentric circles.}

\item{text_size}{Text size for axes.}

\item{text_opacity}{Text opacity for axes.}

\item{ellipse_opacity}{Opacity of confidence ellipses.}

\item{overlay_parameter_info}{Whether or not to show amplitude/phase
estimates.}

\item{fill_colors_check}{\code{!missing(fill_colors)}.}

\item{xlims_check}{\code{TRUE} if limits are unspecified.}

\item{xlims}{User-specified limits.}

\item{ylims_check}{\code{TRUE} if limits are unspecified.}

\item{ylims}{User-specified limits.}

\item{quietly}{Send possibly unnecessary messages to console.}

\item{overlay_start}{background axes positioning.}

\item{fill_colors}{Colours for ellipses.}

\item{zoom_origin}{Zoom position if used.}

\item{ranef_plot}{A \code{character} naming a random-effect grouping
variable in \code{x}, whose level estimates (from
\code{ranef_estimates()}) will be overlaid as points. \code{NULL}
(default) for no overlay.}

\item{ranef_ci}{\code{NULL} for no confidence ellipse, or a single
number giving the confidence level at which to draw a bootstrap
confidence ellipse around each \code{ranef_plot} point.}
}
\value{
A \code{ggplot2} object.
}
\description{
create a (sub) polar plot (iterated over for each component)
}
\examples{
model <- cglmm(
  vit_d ~ amp_acro(time_col = time, group = "X", period = 12),
  data = vitamind
)
GLMMcosinor:::sub_ggplot.cglmm.polar(
  comp = 1,
  x = model,
  sum = summary(model),
  direction = 1,
  offset = 0,
  radial_units = "period",
  grid_angle_segments = 8,
  n_breaks = 5,
  zoom = FALSE,
  circle_linetype = "dotted",
  text_size = 3.5,
  text_opacity = 1,
  ellipse_opacity = 0.3,
  overlay_parameter_info = FALSE,
  fill_colors_check = FALSE,
  xlims_check = FALSE,
  ylims_check = FALSE,
  quietly = TRUE,
  overlay_start = 0,
  zoom_origin = FALSE
)

}
