% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cosinor.R
\name{test_cosinor}
\alias{test_cosinor}
\title{Test for differences in a cosinor model}
\usage{
test_cosinor(
  x,
  x_str,
  param = c("amp", "acr"),
  comparison_A = 0,
  comparison_B = 1,
  comparison_type = c("levels", "components"),
  component_index = 1,
  level_index = 0,
  ci_level = 0.95
)
}
\arguments{
\item{x}{An \code{cosinor.glmm} object.}

\item{x_str}{A \code{character}. The name of the grouping variable within
which differences in the selected cosinor characteristic (amplitude or
acrophase) will be tested.}

\item{param}{A \code{character}. Either \code{"amp"} or \code{"acr"} for
testing differences in amplitude or acrophase, respectively.}

\item{comparison_A}{An \code{integer}. Refers to the level (within the
grouping variable) or component number that is to act as the reference group
for the comparison.}

\item{comparison_B}{An \code{integer}. Refers to the level (within the
grouping variable) or component number that is to act as the comparator group
for the comparison.}

\item{comparison_type}{A \code{character}. Indicates whether the comparison
to be performed is to be between levels of a grouping variable
(\code{"levels"}) or indices of components in a multiple-component cosinor
model (\code{"components"}).}

\item{component_index}{An \code{integer}. If
\code{comparison_type = "levels"}, \code{component_index} indicates which
component is being compared between the levels of the grouping variable.}

\item{level_index}{An \code{integer}. If
\code{comparison_type = "components"}, \code{level_index} indicates which
level of the grouping variable is being used for the comparison between
components.}

\item{ci_level}{The level for calculated confidence intervals. Defaults to
\code{0.95}.}
}
\value{
Returns a \code{test_cosinor} object.
}
\description{
Given a time variable and optional covariates, generate inference a cosinor
fit. For the covariate named (or vector of covariates), this function
performs a Wald test comparing the group with covariates equal to 1 to the
group with covariates equal to 0. This may not be the desired result for
continuous covariates.
}
\examples{
mod_grouped <- cosinor.glmm(Y ~ X + amp_acro(time,
  group = "X",
  n_components = 1,
  period = 12
), data = vitamind)

test_cosinor(mod_grouped, "X", "amp")

data_2_component <- simulate_cosinor(
  n = 10000,
  mesor = 5,
  amp = c(2, 5),
  acro = c(0, pi),
  family = "gaussian",
  n_components = 2,
  period = c(10, 12)
)

mod_2_component <- cosinor.glmm(
  Y ~ amp_acro(times, n_components = 2, period = c(10,12)),
  data = data_2_component
)

# test_cosinor(mod_2_component, param = "amp", comparison_type = "components")

}
