% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.cglmm}
\alias{autoplot.cglmm}
\title{Plot a cosinor model}
\usage{
\method{autoplot}{cglmm}(
  object,
  ci_level = 0.95,
  x_str,
  type = "response",
  xlims,
  pred.length.out,
  points_per_min_cycle_length = 20,
  superimpose.data = FALSE,
  data_opacity = 0.3,
  predict.ribbon = TRUE,
  ranef_plot = NULL,
  ...
)
}
\arguments{
\item{object}{An \code{cglmm} object.}

\item{ci_level}{The level for calculated confidence intervals. Defaults to
\code{0.95}.}

\item{x_str}{A \code{character} vector naming variable(s) to be plotted.
Default has no value and plots all groups.}

\item{type}{A \code{character} that will be passed as an argument to
\code{predict.cglmm()}, specifying the type of prediction
(e.g, "response", or "link"). See \code{?glmmTMB::predict.glmmTMB} for full
list of possible inputs.}

\item{xlims}{A vector of length two containing the limits for the x-axis.}

\item{pred.length.out}{An integer value that specifies the number of
predicted data points. The larger the value, the more smooth the fitted line
will appear. If missing, uses \code{points_per_min_cycle_length} to generate
a sensible default value.}

\item{points_per_min_cycle_length}{Used to determine the number of samples
to create plot if \code{pred.length.out} is missing.}

\item{superimpose.data}{A \code{logical}. If \code{TRUE}, data from the
original data used to fit the model (\code{object}) will be superimposed
over the predicted fit.}

\item{data_opacity}{A number between 0 and 1 inclusive that controls the
opacity of the superimposed data. (Used as the \code{alpha} when calling
\code{ggplot2::geom_point()}).}

\item{predict.ribbon}{A code{logical}. If \code{TRUE}, a prediction interval
is plotted.}

\item{ranef_plot}{Specify the random effects variables that you wish to plot.
If not specified, only the fixed effects will be visualised.
\code{points_per_min_cycle_length} is the number of points plotted per the
minimum cycle length (period) of all cosinor components in the model.}

\item{...}{Additional, ignored arguments.}
}
\value{
Returns a \code{ggplot} object.
}
\description{
Given a cglmm model fit, generate a plot of the data with the fitted
values.
Optionally allows for plotting by covariates
}
\examples{
model <- cglmm(
  vit_d ~ X + amp_acro(time, group = "X", period = 12),
  data = vitamind
)
autoplot(model, x_str = "X")
}
