% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.cosinor.glmm}
\alias{autoplot.cosinor.glmm}
\title{Plot a cosinor model}
\usage{
\method{autoplot}{cosinor.glmm}(
  object,
  ci_level = 0.95,
  x_str,
  type = "response",
  xlims,
  pred.length.out,
  points_per_min_cycle_length = 20,
  superimpose.data = FALSE,
  data_opacity = 0.3,
  predict.ribbon = TRUE,
  ...
)
}
\arguments{
\item{object}{An \code{cosinor.glmm} object.}

\item{ci_level}{The level for calculated confidence intervals. Defaults to 0.95.}

\item{x_str}{Character vector naming covariate(s) to be plotted. Default has no value and plots all groups}

\item{type}{Character that will be passed as an argument to the predict.cosinor.glmm() function, specifying the type of prediction (e.g, "response", or "link")}

\item{xlims}{A vector of length two containing the lower and upper x limit to be plotted}

\item{pred.length.out}{An integer value that specifies the number of predicted datapoints. The larger the value, the more smooth the fit will appear}

\item{points_per_min_cycle_length}{Number of datapoints for the smallest period}

\item{superimpose.data}{A logical argument (TRUE or FALSE). If TRUE, data from the original cosinor.glmm() object will be superimposed over the predicted fit}

\item{data_opacity}{A number bewteen 0 and 1 inclusive that controls the opacity of the superimposed data}

\item{predict.ribbon}{A logical argument (TRUE or FALSE). If TRUE, a prediction interval is plotted}

\item{...}{Additional, ignored arguments.}
}
\value{
Returns a \code{ggplot} graphics object.
}
\description{
Given a cosinor.glmm model fit, generate a plot of the data with the fitted values.
Optionally allows for plotting by covariates
}
\examples{

model <- cosinor.glmm(Y ~ X + amp_acro(time, group = "X", period = 12), data = vitamind)
autoplot(model, x_str = "X")

}
