% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processor.R
\name{fit_model_and_process}
\alias{fit_model_and_process}
\title{Fit the cosinorGLMM model using the output from update_formula_and_data() and
a new formula}
\usage{
fit_model_and_process(obj, formula, ...)
}
\arguments{
\item{obj}{output from \code{update_formula_and_data()}}

\item{formula}{new formula to use when fitting model (maybe with random effects)}

\item{...}{optional additional arguments passed to \code{glmmTMB::glmmTMB()}}
}
\value{
Returns a fitted cosinor model as a \code{cosinor.glmm} object.
}
\description{
Fit the cosinorGLMM model using the output from update_formula_and_data() and
a new formula
}
\examples{
vitamind2 <- vitamind
vitamind2$patient <- sample(LETTERS[1:5], size = nrow(vitamind2), replace = TRUE)
updated_df_and_formula <- update_formula_and_data(
  data = vitamind2,
  formula = Y ~ X + amp_acro(time,
    group = "X",
    period = 12
  )
)
res <- fit_model_and_process(
  updated_df_and_formula,
  update.formula(updated_df_and_formula$newformula, . ~ . + (1 | patient))
)

}
