% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cosinor.glmm.R
\name{ggplot.cosinor.glmm}
\alias{ggplot.cosinor.glmm}
\title{Plot a cosinor model}
\usage{
\method{ggplot}{cosinor.glmm}(
  object,
  x_str = NULL,
  type = "response",
  xlims,
  pred.length.out = 200,
  superimpose.data = FALSE,
  data_opacity = 0.3,
  predict.ribbon = TRUE
)
}
\arguments{
\item{object}{An object of class \code{cosinor.glmm}}

\item{x_str}{Character vector naming the covariate(s) to be plotted. May be NULL to plot overall curve}
}
\description{
Given a cosinor.glmm model fit, generate a plot of the data with the fitted values.
Optionally allows for plotting by covariates
}
\examples{

object <- cosinor.glmm(Y ~ X + amp.acro(time, group = "X"), data = vitamind)
ggplot.cosinor.glmm(object, x_str = "X")

}
