% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMMcosinor-package.R
\docType{package}
\name{GLMMcosinor-package}
\alias{GLMMcosinor}
\alias{GLMMcosinor-package}
\title{GLMMcosinor: Fit a cosinor model using a generalised mixed modelling framework}
\description{
Fit the cosinor model using the glmmTMB framework, allowing for a wide range of link functions and capacity of heirachical structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RWParsons/GLMMcosinor}
}

}
\author{
\strong{Maintainer}: Rex Parsons \email{rex.parsons94@gmail.com} (\href{https://orcid.org/0000-0002-6053-8174}{ORCID})

Authors:
\itemize{
  \item Oliver Jayasinghe
}

}
\keyword{internal}
