% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMMcosinor-package.R
\docType{package}
\name{GLMMcosinor-package}
\alias{GLMMcosinor}
\alias{GLMMcosinor-package}
\title{GLMMcosinor: A package for fitting generalized linear mixed-effects
models with cosinor terms}
\description{
To learn more about how to use GLMMcosinor, see the vignettes at the
\href{https://rwparsons.github.io/GLMMcosinor/}{website} or using
the following code:

\code{browseVignettes(package = "GLMMcosinor")}

This package provides functions for fitting generalized linear mixed-effects
models with cosinor terms. Cosinor terms can be used to model periodic
variations in time series data, and are particularly useful in chronobiology
and sleep research. The package uses the flexible GLMMtmb framework that
allows users to specify a wide range of data families, and provides tools
for visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RWParsons/GLMMcosinor}
  \item \url{https://rwparsons.github.io/GLMMcosinor/}
  \item Report bugs at \url{https://github.com/RWParsons/GLMMcosinor/issues}
}

}
\author{
\strong{Maintainer}: Rex Parsons \email{rex.parsons94@gmail.com} (\href{https://orcid.org/0000-0002-6053-8174}{ORCID})

Authors:
\itemize{
  \item Oliver Jayasinghe
  \item Nicole White (\href{https://orcid.org/0000-0002-9292-0773}{ORCID})
}

}
\keyword{internal}
