% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cosinor.glmm.R
\name{polar_plot}
\alias{polar_plot}
\title{Generates a polar plot with elliptical confidence intervals}
\usage{
polar_plot(
  x,
  contour_interval = 1,
  make_cowplot = TRUE,
  component_index = 1,
  grid_angle_segments = 8,
  radial_units = "radians",
  clockwise = FALSE,
  text_size = 3,
  text_opacity = 0.5,
  fill_colours,
  ellipse_opacity = 0.3,
  circle_linetype = "dotted",
  start = "right",
  view = "full",
  overlay_parameter_info = FALSE,
  quietly = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{cosinor.glmm}}

\item{contour_interval}{The distance bewteen adjacent circular contours in the background of the polar plot}

\item{make_cowplot}{A logical argument. If TRUE, plots polar plots for each component and displays the results as a single output with several plots. If make_cowplot is TRUE, specifying component_index is redundant}

\item{component_index}{A number that corresponds to a particular component from the cosinor.glmm() object that will be used to create polar plot. If make_cowplot is FALSE, then component_index controls which component is plotted}

\item{grid_angle_segments}{An integer that determines the total number of segments in the background of the polar plot. For example, a value of 4 will create quadrants around the origin.}

\item{radial_units}{A string controlling the angle units. Valid arguments are: 'radians', 'degrees', or 'period'. Radians plots from 0 to 2*pi; degrees plots from 0 to 360, and period plots from 0 to the maximum period in the component}

\item{clockwise}{A logical argument. If TRUE, the angles increase in a clockwise fashion}

\item{text_size}{A number controlling the size of the text labels}

\item{text_opacity}{A number between 0 and 1 inclusive which determines the opacity of the text labels}

\item{fill_colours}{A vector containing colours (expressed as strings) that will be used to delineate levels within a group. If the model has components with different number of levels per factor, the length of this input should match the greatest number of levels. If not, or if the number of levels exceeds the length of the default argument (8), colours are generated using rainbow()}

\item{ellipse_opacity}{A number between 0 and 1 inclusive which determines the opacity of the confidence ellipses}

\item{circle_linetype}{A string which determines the linetype of the radial circles in background of the polar plot. See ?linetype for more details}

\item{start}{A character, either "right", "left", "top", or "bottom" which determines where angle 0 is located. If start = "top", and clockwise = TRUE, the angle will rotate clockwise, starting at the '12' position on a clock}

\item{view}{A character, either "full", "zoom", or "zoom_origin" which controls the view of the plots. "full" maintains a full view of the polar plot, including the background radial circles. "zoom" finds the minimum viewwindow which contains all confidence ellipses. "zoom_origin" zooms into the confidence ellipses (like "zoom"), but also keeps the origin within frame}

\item{overlay_parameter_info}{A logical argument. If TRUE, more information about the acrophase and amplitude are overlayed onto the polar plots.}

\item{quietly}{Analagous to verbose, this logical argument controls whether messages are displayed in the console.}

\item{...}{Additional, ignored arguments.}
}
\description{
Generates a polar plot with elliptical confidence intervals
}
\examples{
object <- cosinor.glmm(Y ~ X + amp.acro(time, group = "X"), data = vitamind)
polar_plot(object)
}
