% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cosinor.glmm.R
\name{predict.cosinor.glmm}
\alias{predict.cosinor.glmm}
\title{Predict from a cosinor model}
\usage{
\method{predict}{cosinor.glmm}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{cosinor.glmm}}

\item{newdata}{Optional new data}

\item{...}{other arguments passed to \code{glmmTMB:::predict.glmmTMB}}
}
\value{
Returns predicted values from the cosinor model.
}
\description{
Given a time variable and optional covariates, generate predicted values from
a cosinor fit. Default prediction is the mean value, optionally can predict
at a given month
}
\examples{

fit <- cosinor.glmm(Y ~ X + amp_acro(time,
  group = "X",
  n_components = 1,
  period = 12
), data = vitamind)
predict(fit)

}
