% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.glmm.R
\name{print.cosinor.glmm}
\alias{print.cosinor.glmm}
\title{Print a \code{cosinor.glmm} model.}
\usage{
\method{print}{cosinor.glmm}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{A \code{cosinor.glmm} object.}

\item{digits}{Controls the number of digits displayed in the summary output.}

\item{...}{Additional, ignored arguments.}
}
\value{
\code{print(x)} returns \code{x} invisibly.
}
\description{
Given an outcome and time variable, fit the cosinor model with optional covariate effects.
}
