% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.glmm.R
\name{print.cosinor.glmm}
\alias{print.cosinor.glmm}
\title{Print cosinor model}
\usage{
\method{print}{cosinor.glmm}(x, ...)
}
\arguments{
\item{x}{cosinor.glmm object}

\item{...}{passed to summary}
}
\description{
Given an outcome and time variable, fit the cosinor model with optional covariate effects.
}
