% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cosinor.glmm}
\alias{print.cosinor.glmm}
\title{Print a brief summary of the \code{cosinor.glmm} model.}
\usage{
\method{print}{cosinor.glmm}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{A \code{cosinor.glmm} object.}

\item{digits}{Controls the number of digits displayed in the summary output.}

\item{...}{Additional, ignored arguments.}
}
\value{
\code{print(x)} returns \code{x} invisibly.
}
\description{
Print a brief summary of the \code{cosinor.glmm} model.
}
