% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cosinor.glmm.R
\name{print.summary.cosinor.glmm}
\alias{print.summary.cosinor.glmm}
\title{Print the summary of a cosinor model}
\usage{
\method{print}{summary.cosinor.glmm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.cosinor.glmm}}

\item{...}{Currently unused}
}
\description{
Print the summary of a cosinor model
}
\examples{

fit <- cosinor.glmm(Y ~ time(time) + X + amp.acro(X), data = vitamind)
summary(fit)

}
