% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cosinor.R
\name{test_cosinor}
\alias{test_cosinor}
\title{Test for differences in a cosinor model}
\usage{
test_cosinor(
  x,
  x_str,
  param = c("amp", "acr"),
  comparison_A = 0,
  comparison_B = 1,
  comparison_type = "levels",
  component_index = 1,
  level_index = 0,
  ci_level = 0.95
)
}
\arguments{
\item{x}{An object of class \code{cosinor.glmm}}

\item{x_str}{Character naming the covariate whose amplitude/acrophase will be tested}

\item{param}{Character string naming the parameter to test, either "amp" for
amplitude or "acr" for acrophase}

\item{comparison_A}{A number, referring to the level (within a group) or component number that is to be compared to comparison_B}

\item{comparison_B}{A number, referring to the level (within a group) or component number that is to be compared to comparison_A}

\item{comparison_type}{A string that is either: "levels" (default), or "components". If "levels", then comparison_A and comparison_B will refer to the two levels that are being compared. If comparison_type = "components", then comparison_A component will be compared to comparison_B component}

\item{component_index}{If comparison_type = "levels", this controls which single component the levels are being compared to. Note that component_index must be an integer, and must refer to a component within the model}

\item{level_index}{If comparison_type = "components", this controls which single level the components are being compared to. Note that level_index must be an integer, and must refer to a level within the model}

\item{ci_level}{The level for calculated confidence intervals. Defaults to 0.95.}
}
\value{
Returns a \code{test_cosinor} object.
}
\description{
Given a time variable and optional covariates, generate inference a cosinor
fit. For the covariate named (or vector of covariates), this function
performs a Wald test comparing the group with covariates equal to 1 to the
group with covariates equal to 0. This may not be the desired result for
continuous covariates.
}
\examples{

fit <- cosinor.glmm(Y ~ X + amp.acro(time,
  group = "X",
  n_components = 1,
  period = 12
), data = vitamind)
test_cosinor(fit, "X", "amp")

}
