% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{update_formula_and_data}
\alias{update_formula_and_data}
\title{Update data and formula for fitting cosinor.glmm model}
\usage{
update_formula_and_data(data, formula, family = "gaussian", ...)
}
\arguments{
\item{data}{input data for fitting cosinor.glmm model.}

\item{formula}{model formula, specified by user including \code{amp.acro()}.}

\item{family}{the model family.}

\item{...}{optional additional arguments.}
}
\value{
Returns a \code{list}.
}
\description{
Update data and formula for fitting cosinor.glmm model
}
